refactor(core): extract Publisher.ts scaffold (behavior-identical)
Carves runtime.deployContract (~250 LoC closure inside initRuntime) out
into src/core/Publisher.ts. runtime.deployContract becomes a 9-line
orchestrator that constructs a Publisher and forwards moduleName /
config / account / packageDir. The class is marked @internal — not
part of the published movehat API surface (revisit at M5).

Pure extraction: zero logic changes. The destructive Move.toml regex
rewrite (bug #38), the shared ~/.aptos/config.yaml mutation (bug #37),
and the SIGINT-vulnerable window between yaml write and finally restore
(bug #36) all still exist verbatim — they get addressed in commits 2-4
of this PR.

Dynamic `await import('fs' / 'path' / 'os' / 'js-yaml' / './shell.js' /
'./compile.js')` calls inside the closure are now static imports at the
top of Publisher.ts. Equivalent runtime semantics because Publisher.ts
is only loaded when runtime.ts is, and runtime.ts is loaded whenever
deployContract could be called.

PublishInput carries the resolved Account (not just privateKey) because
the yaml profile write needs both publicKey.toString() and
accountAddress.toString() — reconstructing those from a raw key would
require an extra SDK call per deploy.

181/181 unit tests pass unchanged.

Refs #79, refs #68.
